Update 3/0000/0072.txt
Made some clarifications to existing translations that were unclear and complained about in issues.

Translated the majority of the remaining lines.
なぞのオンナが ここ  was left untranslated as the script itself makes it unclear who the speaker is referring to as the "Mysterious Woman" or 
"Mysterious Girl" Basically, she is saying いいね！　Great! あなた,なぞのおんあ　ここ！ you, mysterious girl, over here! However, it also seems strange that the speaker would refer to the player as a "mysterious girl" when she just mentioned that you have a battle recorder. Just not sure of the context surrounding this scene here.

Puoba! could possibly be Pansear talking in Pokemon. Baoppu is the Japanese name for Pansear, so this line could be something like: "ear An Pan!" or "Pan! Sear Sear!" or words to that effect. But again, it is not clear if Pansear is actually talking as it is not marked as such, so "Puoba" it is until someone can clear that up.

But I did not see a separate line if the player is a boy and that makes it difficult to be sure that the speaker is referring to the player or some other person.